refactor: create utils to jwt example
